test(browser): Remove webpack build tests (#12967)
Removes `packages/browser/test/package/npm-build.js`.

`packages/browser/test/package/npm-build.js` contained three tests:

1. Test against `tslib_1__default` appearing in the bundle. I added a
browser integration test to replicate this
2. Test that adding a bundled package would work, our integration + e2e
tests pretty much cover this
3. Test that you can add bundle packages multiple times - our
integration + e2e tests also cover this

I also migrated `packages/utils/test/types/index.js` to use jest instead
of a plain script.
