+++ .github/workflows/build.yml
@@ -779,33 +779,6 @@ jobs:
           name: playwright-traces
           path: dev-packages/browser-integration-tests/test-results
 
-  job_browser_build_tests:
-    name: Browser Build Tests
-    needs: [job_get_metadata, job_build]
-    runs-on: ubuntu-20.04
-    timeout-minutes: 5
-    steps:
-      - name: Check out current commit (${{ needs.job_get_metadata.outputs.commit_label }})
-        uses: actions/checkout@v4
-        with:
-          ref: ${{ env.HEAD_COMMIT }}
-      - name: Set up Node
-        uses: actions/setup-node@v4
-        with:
-          node-version-file: 'package.json'
-      - name: Restore caches
-        uses: ./.github/actions/restore-cache
-        env:
-          DEPENDENCY_CACHE_KEY: ${{ needs.job_build.outputs.dependency_cache_key }}
-      - name: Run browser build tests
-        run: |
-          cd packages/browser
-          yarn test:package
-      - name: Run utils build tests
-        run: |
-          cd packages/utils
-          yarn test:package
-
   job_check_for_faulty_dts:
     name: Check for faulty .d.ts files
     needs: [job_get_metadata, job_build]
@@ -1317,7 +1290,6 @@ jobs:
       [
         job_build,
         job_compile_bindings_profiling_node,
-        job_browser_build_tests,
         job_browser_unit_tests,
         job_bun_unit_tests,
         job_deno_unit_tests,
