+++ packages/browser/package.json
@@ -49,8 +49,7 @@
   },
   "devDependencies": {
     "@sentry-internal/integration-shims": "8.18.0",
-    "fake-indexeddb": "^4.0.1",
-    "webpack": "^4.47.0"
+    "fake-indexeddb": "^4.0.1"
   },
   "scripts": {
     "build": "run-p build:transpile build:bundle build:types",
@@ -73,7 +72,6 @@
     "size:check": "cat build/bundles/bundle.min.js | gzip -9 | wc -c | awk '{$1=$1/1024; print \"ES2017: \",$1,\"kB\";}'",
     "test": "yarn test:unit",
     "test:unit": "jest",
-    "test:package": "node test/package/npm-build.js && rm test/package/tmp.js",
     "test:unit:watch": "jest --watch",
     "yalc:publish": "yalc publish --push --sig"
   },
