+++ packages/utils/package.json
@@ -63,7 +63,6 @@
     "lint": "eslint . --format stylish",
     "test": "jest",
     "test:watch": "jest --watch",
-    "test:package": "node test/types/index.js",
     "version": "node ../../scripts/versionbump.js src/version.ts",
     "yalc:publish": "yalc publish --push --sig"
   },
